requirements.txt
asgiref==3.6.0
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
django-admin-extra-buttons==1.5.6
django-admin-list-filter-dropdown==1.0.3
django-colorfield==0.8.0
django-filter==22.1
django-tinymce==3.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
idna==3.4
Markdown==3.4.1
mysqlclient==2.1.1
Pillow==9.4.0
psycopg2-binary==2.9.3
pycodestyle==2.10.0
PyJWT==1.7.1
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.15
whitenoise==5.3.0
